refactor(test): group remaining messaging and policy coverage
