+++ test/runtime/messaging/voice-gateway-integration.test.ts
@@ -8,11 +8,11 @@ import path from "node:path";
 
 import { afterEach, describe, expect, it } from "vitest";
 
-import { createVoiceGatewayServer } from "../src/lib/adapters/http/voice-gateway-server";
-import type { AgentTurnClient, AgentTurnEvent } from "../src/lib/voice-gateway/contracts";
-import { readPrivateBearerDescriptors } from "../src/lib/voice-gateway/credential-file";
-import { OpenClawVoiceClient } from "../src/lib/voice-gateway/openclaw-client";
-import { VoiceSessionService } from "../src/lib/voice-gateway/session-service";
+import { createVoiceGatewayServer } from "../../../src/lib/adapters/http/voice-gateway-server";
+import type { AgentTurnClient, AgentTurnEvent } from "../../../src/lib/voice-gateway/contracts";
+import { readPrivateBearerDescriptors } from "../../../src/lib/voice-gateway/credential-file";
+import { OpenClawVoiceClient } from "../../../src/lib/voice-gateway/openclaw-client";
+import { VoiceSessionService } from "../../../src/lib/voice-gateway/session-service";
 import { PinnedOpenClawGateway } from "./fixtures/voice-gateway/pinned-openclaw-gateway";
 import { PinnedVoiceRuntimeAdapter } from "./fixtures/voice-gateway/pinned-runtime-adapter";
