+++ test/runtime/messaging/wechat-diagnostics.test.ts
