+++ test/runtime/policy/policies-teams.test.ts
@@ -8,11 +8,11 @@ import os from "node:os";
 import path from "node:path";
 import { describe, expect, it } from "vitest";
 
-import * as policies from "../src/lib/policy";
+import * as policies from "../../../src/lib/policy";
 
 const requireForTest = createRequire(import.meta.url);
 const YAML = requireForTest("yaml");
-const REPO_ROOT = path.join(import.meta.dirname, "..");
+const REPO_ROOT = path.join(import.meta.dirname, "../../..");
 const POLICIES_PATH = JSON.stringify(path.join(REPO_ROOT, "src", "lib", "policy", "index.ts"));
 const REGISTRY_PATH = JSON.stringify(path.join(REPO_ROOT, "src", "lib", "state", "registry.ts"));
