+++ test/runtime/policy/policies.test.ts
@@ -10,7 +10,7 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 const requireForTest = createRequire(import.meta.url);
 const YAML = requireForTest("yaml");
-const REPO_ROOT = path.join(import.meta.dirname, "..");
+const REPO_ROOT = path.join(import.meta.dirname, "../../..");
 const policies = requireForTest(
   path.join(REPO_ROOT, "src", "lib", "policy", "index.ts"),
 ) as typeof import("../src/lib/policy");
